pyproject.toml
[project]
name = "clmusdiffv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.2.0",
    "diffusers>=0.35.1",
    "dotenv>=0.9.9",
    "ipykernel>=7.0.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
    "torch-tb-profiler>=0.4.3",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
    "xformers>=0.0.32.post2",
]
